+++ tools/linter_lib/custom_check.py
@@ -10,6 +10,7 @@ from zulint.custom_rules import Rule, RuleList
 
 FILES_WITH_LEGACY_SUBJECT = {
     # This basically requires a big DB migration:
+    "zerver/lib/message.py",
     "zerver/lib/topic.py",
     "zerver/lib/topic_sqlalchemy.py",
     # This is for backward compatibility.