+++ web/src/stream_types.ts
@@ -28,6 +28,7 @@ export const stream_schema = z.object({
         MODERATORS: StreamPostPolicy.MODERATORS,
     }),
     can_remove_subscribers_group: z.number(),
+    can_access_stream_topics_group: z.number(),
 });
 
 export const stream_specific_notification_settings_schema = z.object({