+++ zerver/lib/message.py
@@ -27,7 +27,13 @@ from zerver.lib.stream_subscription import (
     num_subscribers_for_stream_id,
 )
 from zerver.lib.streams import can_access_stream_history, get_web_public_streams_queryset
-from zerver.lib.topic import MESSAGE__TOPIC, TOPIC_NAME, messages_for_topic
+from zerver.lib.topic import (
+    MESSAGE__TOPIC,
+    TOPIC_NAME,
+    check_access_based_on_can_access_stream_topics_group,
+    get_topic_creator_user_id,
+    messages_for_topic,
+)
 from zerver.lib.types import UserDisplayRecipient
 from zerver.lib.user_groups import user_has_permission_for_group_setting
 from zerver.lib.user_topics import build_get_topic_visibility_policy, get_topic_visibility_policy
@@ -47,6 +53,7 @@ from zerver.models.constants import MAX_TOPIC_NAME_LENGTH
 from zerver.models.groups import SystemGroups
 from zerver.models.messages import get_usermessage_by_message_id
 from zerver.models.realms import WildcardMentionPolicyEnum
+from zerver.models.streams import get_stream_by_id_in_realm
 from zerver.models.users import is_cross_realm_bot_email
 
 
@@ -171,6 +178,7 @@ class SendMessageRequest:
     disable_external_notifications: bool = False
     automatic_new_visibility_policy: int | None = None
     recipients_for_user_creation_events: dict[UserProfile, set[int]] | None = None
+    is_support_stream: bool | None = None
 
 
 # We won't try to fetch more unread message IDs from the database than
@@ -228,6 +236,19 @@ def messages_for_ids(
 
     for message_id in message_ids:
         msg_dict = message_dicts[message_id]
+        if user_profile and msg_dict["recipient_type"] == Recipient.STREAM:
+            stream_id = msg_dict["recipient_type_id"]
+            stream = get_stream_by_id_in_realm(stream_id, realm)
+            if (
+                stream.is_support_stream()
+                and not check_access_based_on_can_access_stream_topics_group(user_profile, stream)
+            ):
+                topic_creator_user_id = get_topic_creator_user_id(
+                    msg_dict["sender_realm_id"], msg_dict["recipient_id"], msg_dict["subject"]
+                )
+                if topic_creator_user_id and user_profile.id != topic_creator_user_id:
+                    continue
+
         flags = user_message_flags[message_id]
         # TODO/compatibility: The `wildcard_mentioned` flag was deprecated in favor of
         # the `stream_wildcard_mentioned` and `topic_wildcard_mentioned` flags.  The
@@ -377,7 +398,9 @@ def has_message_access(
         return has_user_message()
 
     if stream is None:
-        stream = Stream.objects.get(id=message.recipient.type_id)
+        stream = Stream.objects.select_related(
+            "can_access_stream_topics_group__named_user_group"
+        ).get(id=message.recipient.type_id)
     else:
         assert stream.recipient_id == message.recipient_id
 
@@ -392,6 +415,15 @@ def has_message_access(
         return Subscription.objects.filter(
             user_profile=user_profile, active=True, recipient=message.recipient
         ).exists()
+
+    if stream.is_support_stream() and not check_access_based_on_can_access_stream_topics_group(
+        user_profile, stream
+    ):
+        topic_creator_user_id = get_topic_creator_user_id(
+            message.get_realm().id, message.recipient_id, message.topic_name()
+        )
+        if topic_creator_user_id and user_profile.id != topic_creator_user_id:
+            return False
 
     if stream.is_public() and user_profile.can_access_public_streams():
         return True
@@ -469,6 +501,20 @@ def bulk_access_stream_messages_query(
     """
 
     messages = messages.filter(realm_id=user_profile.realm_id, recipient_id=stream.recipient_id)
+
+    first_message = messages.first()
+    if (
+        stream.is_support_stream()
+        and not check_access_based_on_can_access_stream_topics_group(user_profile, stream)
+        and first_message is not None
+    ):
+        topic_creator_user_id = get_topic_creator_user_id(
+            first_message.get_realm().id,
+            first_message.recipient_id,
+            first_message.topic_name(),
+        )
+        if topic_creator_user_id and user_profile.id != topic_creator_user_id:
+            return Message.objects.none()
 
     if stream.is_public() and user_profile.can_access_public_streams():
         return messages