+++ zerver/tests/test_message_topics.py
@@ -2,10 +2,11 @@ from unittest import mock
 
 from django.utils.timezone import now as timezone_now
 
-from zerver.actions.streams import do_change_stream_permission
+from zerver.actions.streams import do_change_stream_group_based_setting, do_change_stream_permission
 from zerver.lib.test_classes import ZulipTestCase
-from zerver.models import Message, UserMessage
+from zerver.models import Message, NamedUserGroup, Stream, UserMessage
 from zerver.models.clients import get_client
+from zerver.models.groups import SystemGroups
 from zerver.models.realms import get_realm
 from zerver.models.streams import get_stream
 
@@ -36,15 +37,14 @@ class TopicHistoryTest(ZulipTestCase):
         stream_name = "Verona"
 
         stream = get_stream(stream_name, user_profile.realm)
-        recipient = stream.recipient
 
-        def create_test_message(topic_name: str) -> int:
+        def create_test_message(topic_name: str, stream: Stream) -> int:
             # TODO: Clean this up to send messages the normal way.
 
             hamlet = self.example_user("hamlet")
             message = Message(
                 sender=hamlet,
-                recipient=recipient,
+                recipient=stream.recipient,
                 realm=stream.realm,
                 content="whatever",
                 date_sent=timezone_now(),
@@ -64,19 +64,17 @@ class TopicHistoryTest(ZulipTestCase):
         # our most recent topics are topic0, topic1, topic2
 
         # Create old messages with strange spellings.
-        create_test_message("topic2")
-        create_test_message("toPIc1")
-        create_test_message("toPIc0")
-        create_test_message("topic2")
-        create_test_message("topic2")
-        create_test_message("Topic2")
+        create_test_message("topic2", stream)
+        create_test_message("toPIc1", stream)
+        create_test_message("toPIc0", stream)
+        create_test_message("topic2", stream)
+        create_test_message("topic2", stream)
+        create_test_message("Topic2", stream)
 
         # Create new messages
-        topic2_msg_id = create_test_message("topic2")
-        create_test_message("topic1")
-        create_test_message("topic1")
-        topic1_msg_id = create_test_message("topic1")
-        topic0_msg_id = create_test_message("topic0")
+        topic2_msg_id = create_test_message("topic2", stream)
+        topic1_msg_id = create_test_message("topic1", stream)
+        topic0_msg_id = create_test_message("topic0", stream)
 
         endpoint = f"/json/users/me/{stream.id}/topics"
         result = self.client_get(endpoint, {})
@@ -153,6 +151,36 @@ class TopicHistoryTest(ZulipTestCase):
         self.assertNotIn("topic0", [topic["name"] for topic in history])
         self.assertNotIn("topic1", [topic["name"] for topic in history])
         self.assertNotIn("topic2", [topic["name"] for topic in history])
+
+        # Test for support stream.
+        stream_name = "support_stream"
+        support_stream = self.make_stream(stream_name)
+        self.subscribe(self.example_user("cordelia"), stream_name)
+        self.subscribe(self.example_user("hamlet"), stream_name)
+        support_stream = get_stream(stream_name, user_profile.realm)
+
+        administrators_user_group = NamedUserGroup.objects.get(
+            name=SystemGroups.ADMINISTRATORS,
+            realm=self.example_user("cordelia").realm,
+            is_system_group=True,
+        )
+        do_change_stream_group_based_setting(
+            support_stream,
+            "can_access_stream_topics_group",
+            administrators_user_group,
+            acting_user=self.example_user("cordelia"),
+        )
+
+        create_test_message("topic in support stream", support_stream)
+
+        endpoint = f"/json/users/me/{support_stream.id}/topics"
+        result = self.client_get(endpoint, {})
+        history = self.assert_json_success(result)["topics"]
+        history = history[:3]
+
+        # Cordelia doesn't have these recent history items
+        # as the topic was created by hamlet.
+        self.assertNotIn("topic in support stream", [topic["name"] for topic in history])
 
     def test_bad_stream_id(self) -> None:
         self.login("iago")