+++ zerver/tests/test_users.py
@@ -954,7 +954,7 @@ class QueryCountTest(ZulipTestCase):
         prereg_user = PreregistrationUser.objects.get(email="fred@zulip.com")
 
         with (
-            self.assert_database_query_count(84),
+            self.assert_database_query_count(87),
             self.assert_memcached_count(19),
             self.capture_send_event_calls(expected_num_events=10) as events,
         ):
@@ -2047,8 +2047,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_topic_wildcard_mention=False,
             possible_stream_wildcard_mention=False,
         )
@@ -2087,8 +2088,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_stream_wildcard_mention=False,
         )
         self.assertEqual(info.dm_mention_email_disabled_user_ids, {hamlet.id})
@@ -2101,8 +2103,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_stream_wildcard_mention=False,
         )
         self.assertEqual(info.stream_push_user_ids, {hamlet.id})
@@ -2111,8 +2114,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_stream_wildcard_mention=True,
         )
         self.assertEqual(info.stream_wildcard_mention_user_ids, {hamlet.id, othello.id})
@@ -2126,8 +2130,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_topic_wildcard_mention=True,
             possible_stream_wildcard_mention=False,
         )
@@ -2142,8 +2147,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_topic_wildcard_mention=True,
             possible_stream_wildcard_mention=False,
         )
@@ -2153,8 +2159,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_topic_wildcard_mention=False,
             possible_stream_wildcard_mention=True,
         )
@@ -2164,8 +2171,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_topic_wildcard_mention=True,
             possible_stream_wildcard_mention=True,
         )
@@ -2178,8 +2186,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
         )
         self.assertEqual(info.stream_push_user_ids, set())
 
@@ -2190,8 +2199,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
         )
         self.assertEqual(info.stream_push_user_ids, {hamlet.id})
 
@@ -2210,8 +2220,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
         )
         self.assertEqual(info.stream_push_user_ids, {hamlet.id})
 
@@ -2233,8 +2244,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_stream_wildcard_mention=False,
         )
         self.assertEqual(info.stream_push_user_ids, set())
@@ -2243,8 +2255,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_stream_wildcard_mention=True,
         )
         self.assertEqual(info.stream_push_user_ids, set())
@@ -2258,8 +2271,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=cordelia.id,
+            message_sender=cordelia,
             stream_topic=stream_topic,
+            stream=stream,
             possible_stream_wildcard_mention=True,
         )
         self.assertTrue(hamlet.id in info.muted_sender_user_ids)
@@ -2271,8 +2285,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_stream_wildcard_mention=True,
         )
         self.assertEqual(info.stream_push_user_ids, set())
@@ -2287,8 +2302,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possible_stream_wildcard_mention=True,
         )
         self.assertEqual(info.stream_push_user_ids, set())
@@ -2307,8 +2323,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possibly_mentioned_user_ids={service_bot.id},
         )
         self.assertEqual(
@@ -2331,8 +2348,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
             possibly_mentioned_user_ids={service_bot.id, normal_bot.id},
         )
         self.assertEqual(info.default_bot_user_ids, {normal_bot.id})
@@ -2351,8 +2369,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
         )
         self.assertEqual(info.followed_topic_email_user_ids, {hamlet.id})
         self.assertEqual(info.followed_topic_push_user_ids, {hamlet.id})
@@ -2383,8 +2402,9 @@ class RecipientInfoTest(ZulipTestCase):
         info = get_recipient_info(
             realm_id=realm.id,
             recipient=recipient,
-            sender_id=hamlet.id,
+            message_sender=hamlet,
             stream_topic=stream_topic,
+            stream=stream,
         )
         self.assertEqual(info.followed_topic_email_user_ids, set())
         self.assertEqual(info.followed_topic_push_user_ids, set())
@@ -2406,8 +2426,9 @@ class RecipientInfoTest(ZulipTestCase):
             get_recipient_info(
                 realm_id=realm.id,
                 recipient=invalid_recipient,
-                sender_id=hamlet.id,
+                message_sender=hamlet,
                 stream_topic=stream_topic,
+                stream=stream,
             )