+++ api_docs/changelog.md
@@ -20,6 +20,16 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 10.0
 
+**Feature level 296**
+
+* [`GET /streams`](/api/get-streams),
+  [`POST /users/me/subscriptions`](/api/subscribe),
+  [`PATCH /streams/{stream_id}`](/api/update-stream),
+  [`GET /events`](/api/get-events): Stream objects now
+  include the `can_access_stream_topics_group` integer,
+  which specifies the ID of the user group that has
+  access to all the topics in a stream.
+
 **Feature level 295**
 
 * [`GET /export/realm/consents`](/api/get-realm-export-consents): Added